factor out updating of corgis datasets
